Encode query parameters explicitly as an alist
Directly call `encode-alist` to encode the value of the `:parameters`
argument, instead of dispatching via `encode-element`.

This is because `alist-p` assumes that alists consist entirely of dotted
pairs whose `cdr` is an atom, and thus returns negative if a single
element's `cdr` is in fact a list.
This made it difficult to successfully set a list as the value of an
attribute.